requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.4.1
distlib==0.3.8
Django==5.1.2
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-stubs==3.15.3
filelock==3.15.4
idna==3.10
mypy==1.15.0
mypy-extensions==1.0.0
mysqlclient==2.2.7
pipenv==2024.0.1
platformdirs==4.2.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tomli==2.2.1
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250301
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.26.3
whitenoise==6.9.0
